update order deploy

+++ WMK_BE_Core/WMK_BE_BusinessLogic/Service/Implement/OrderService.cs
@@ -47,7 +47,16 @@ namespace WMK_BE_BusinessLogic.Service.Implement
 			{
 				result.StatusCode = 200;
 				result.Message = "Order List";
-				result.Data = _mapper.Map<List<OrderResponse>>(orders);
+				var orderR = _mapper.Map<List<OrderResponse>>(orders);
+				foreach ( var item in orderR )
+				{
+					var user = await _unitOfWork.UserRepository.GetByIdAsync(item.UserId.ToString());
+					if ( user != null )
+					{
+						item.UserId = user.FirstName + " " + user.LastName;
+					}
+				}
+				result.Data = orderR;
 				return result;
 			}
 			else
@@ -64,11 +73,11 @@ namespace WMK_BE_BusinessLogic.Service.Implement
 			var result = new ResponseObject<List<OrderResponse>>();
 			try
 			{
-                //tim thong tin cua order lien quan toi id nguoi  dung
-                var foundList = _unitOfWork.OrderRepository.Get(x => x.UserId == userId).ToList();
-                if (foundList.Count() == 0)
+				//tim thong tin cua order lien quan toi id nguoi  dung
+				var foundList = _unitOfWork.OrderRepository.Get(x => x.UserId == userId).ToList();
+				if ( foundList.Count() == 0 )
 				{
-					result.StatusCode=500;
+					result.StatusCode = 500;
 					result.Message = "Empty";
 					return result;
 				}
@@ -168,7 +177,7 @@ namespace WMK_BE_BusinessLogic.Service.Implement
 				{
 					await _unitOfWork.CompleteAsync();
 					//change status transaction
-					foreach (var zaloPay in orderExist.Transactions)
+					foreach ( var zaloPay in orderExist.Transactions )
 					{
 						zaloPay.Status = TransactionStatus.PAID;
 						zaloPay.TransactionDate = DateTime.Now;
@@ -230,9 +239,13 @@ namespace WMK_BE_BusinessLogic.Service.Implement
 				return result;
 			}
 			//Tinh so luong trong recipeList
-
+			Random random = new Random();
+			int minValue = 10000000;
+			int maxValue = 99999999;
+			int randomOrderCode = random.Next(minValue , maxValue + 1);
 			//chua tinh duoc total price (dự kiến tính bằng cách nhân quantity với price của từng sản phẩm trong listRecipe nếu là custom hoặc là lấy giá của weeklyPlan nếu là formal
 			var newOrder = _mapper.Map<Order>(model);
+			newOrder.OrderCode = randomOrderCode;
 			//newOrder.TotalPrice = model.TotalPrice * 1000;
 			newOrder.OrderDate = DateTime.Now;
 			newOrder.Status = OrderStatus.Processing;
@@ -252,7 +265,7 @@ namespace WMK_BE_BusinessLogic.Service.Implement
 						zaloPayModel.OrderId = newOrder.Id;
 						zaloPayModel.Amount = newOrder.TotalPrice;
 						var createTransaction = await _transactionService.CreatePaymentZaloPayAsync(zaloPayModel);
-						if ( createTransaction != null && createTransaction.StatusCode == 200 && createTransaction.Data != null)
+						if ( createTransaction != null && createTransaction.StatusCode == 200 && createTransaction.Data != null )
 						{
 							//newOrder.Transactions.Add(createTransaction.Data);
 							await _unitOfWork.CompleteAsync();
@@ -261,8 +274,8 @@ namespace WMK_BE_BusinessLogic.Service.Implement
 							result.Data = newOrder.Id;
 							return result;
 						}
-						if(createTransaction.StatusCode != 200)//code cu la (createTransaction != null)
-                        {
+						if ( createTransaction != null && createTransaction.StatusCode != 200 )//code cu la (createTransaction != null)
+						{
 							// Delete the order if transaction creation fails
 							await _unitOfWork.OrderRepository.DeleteAsync(newOrder.Id.ToString());
 							await _unitOfWork.CompleteAsync();
